Snapshot of work in progress: further work on onm.Store.createComponent...
... onm.Store.createComponent was extended in the last month to support two optional in-parameters. The second, the "construction options" object is cherry-picked to initialize properties of the data component indicated by the onm.Address in-parameter of onm.Store.createComponent. onm before v0.3.4 only shallow-copied to the edge of the component. If the construction options object contained subcomponents, they're ignored. This sucks.

So... I'm slogging through this crap CoffeeScript I wrote. The only saving grace is that I'm getting a lot of practice with Mocha. And, many of the tests can be re-used adapted for planned onm v0.5 re-write of onm.Address.

This commit, I believe, filters out the information required to make construction object data-driven onm.Store.createComponent calls on the subcomponents. More work and tests to get this going...