Notes on test strategy for the component resolution layer of onm.
Still a bit to finish thinking carefully through here. The behavior of the component resolver is hardly given away by its four property input option object... There are quite a few test cases that need coverage here (probably thousands?) I'm trying to figure out which parts of the matrix can be either be ignored (as in don't-care), or tested simply on a single dimension, or two, as opposed to in combination w/all the other dimensions.