Yea right...
This is an outline for baseline test coverage required to ensure that the address resolver the component resolver upon which it depends (thankfully good coverage on named object resolver already...).

This looks like a lot of work. And it is actually. But I think the code basically covers these scenarios and is working. So it will be sort of fun to see how many tests it takes to prove it to a reasonable level of certitude.